refactor(titlebar): extract path helpers to src/shared/path-extractors (#152)
Both sidebar and terminal Titlebars carried duplicate helpers
(extractProjectName, extractWorkgroupName, extractAgentName).
Move them to a single shared module and apply two correctness
fixes while extracting:

- Use lastIndexOf('.ac-new') instead of indexOf to match the
  Rust backend agent_fqn_from_path semantics for nested paths
  (e.g. proj/.ac-new/wg-1/repo-X/.ac-new/wg-2/__agent_Y must
  resolve to the innermost triple). MEDIUM-1 fix.
- Filter empty path segments to handle UNC prefixes, trailing
  slashes, and double slashes deterministically.
- Tighten extractWorkgroupName regex from startsWith('wg-') to
  /^wg-\d+/ so wg- and wg-foo are rejected. LOW-5 fix.

Both Titlebars now import from the shared module ; rtk trailingText
behavior is unchanged in this commit.
