+++ src/sidebar/components/Titlebar.tsx
@@ -2,35 +2,12 @@ import { Component, Show, For, createSignal, createMemo, onMount, onCleanup } fr
 import iconUrl from "../../assets/icon-16.png";
 import { SettingsAPI } from "../../shared/ipc";
 import { isTauri } from "../../shared/platform";
+import { extractProjectName, extractWorkgroupName, extractAgentName } from "../../shared/path-extractors";
 import { terminalStore } from "../../terminal/stores/terminal";
 import type { MainSidebarSide } from "../../shared/types";
 
 declare const __APP_VERSION__: string;
 const APP_VERSION = __APP_VERSION__;
-
-function extractProjectName(workDir: string): string | null {
-  const parts = workDir.replace(/\\/g, '/').split('/');
-  const idx = parts.indexOf('.ac-new');
-  return idx > 0 ? parts[idx - 1] : null;
-}
-
-function extractWorkgroupName(workDir: string): string | null {
-  const parts = workDir.replace(/\\/g, '/').split('/');
-  const idx = parts.indexOf('.ac-new');
-  if (idx < 0 || idx + 1 >= parts.length) return null;
-  const wg = parts[idx + 1];
-  return wg.startsWith('wg-') ? wg.toUpperCase() : null;
-}
-
-function extractAgentName(workDir: string): string | null {
-  const parts = workDir.replace(/\\/g, '/').split('/');
-  const idx = parts.indexOf('.ac-new');
-  if (idx < 0 || idx + 2 >= parts.length) return null;
-  const seg = parts[idx + 2];
-  if (!seg.startsWith('__agent_')) return null;
-  const name = seg.slice('__agent_'.length);
-  return name.length > 0 ? name : null;
-}
 
 const SIDEBAR_WIDTH_PRESETS: Array<{ label: string; width: number }> = [
   { label: "Narrow", width: 200 },
